[core] fix: restore PopoverPosition value export
