+++ packages/core/src/components/index.ts
@@ -78,9 +78,9 @@ export type { IPanel, IPanelProps } from "./panel-stack/panelProps";
 export { PanelStack2, type PanelStack2Props } from "./panel-stack2/panelStack2";
 export type { Panel, PanelProps } from "./panel-stack2/panelTypes";
 export { type PopoverProps, Popover, PopoverInteractionKind } from "./popover/popover";
+export { PopoverPosition } from "./popover/popoverPosition";
 export type {
     DefaultPopoverTargetHTMLProps,
-    PopoverPosition,
     PopoverSharedProps,
     PopoverTargetProps,
     PopoverClickTargetHandlers,
