+++ packages/core/src/components/popover/popoverSharedProps.ts
@@ -18,17 +18,8 @@ import type { Boundary, Modifier, Placement, RootBoundary, StrictModifiers } fro
 import type * as React from "react";
 import type { StrictModifier } from "react-popper";
 
-import { Position, Props } from "../../common";
+import type { Props } from "../../common";
 import type { OverlayableProps } from "../overlay/overlay";
-
-export const PopoverPosition = {
-    ...Position,
-    AUTO: "auto" as "auto",
-    AUTO_END: "auto-end" as "auto-end",
-    AUTO_START: "auto-start" as "auto-start",
-};
-// eslint-disable-next-line @typescript-eslint/no-redeclare
-export type PopoverPosition = (typeof PopoverPosition)[keyof typeof PopoverPosition];
 
 export { Boundary as PopperBoundary, Placement };
 // copied from @popperjs/core, where it is not exported as public
